How to think about Consumers and Producers
