+++ src/Confluent.Kafka/Consumer.cs
@@ -322,37 +322,8 @@ namespace Confluent.Kafka
         public long OutQueueLength
             => consumer.OutQueueLength;
 
-
-        public List<GroupInfo> ListGroups(TimeSpan timeout)
-            => consumer.ListGroups(timeout);
-
-        public List<GroupInfo> ListGroups()
-            => consumer.ListGroups();
-
-
-        public GroupInfo ListGroup(string group, TimeSpan timeout)
-            => consumer.ListGroup(group, timeout);
-
-        public GroupInfo ListGroup(string group)
-            => consumer.ListGroup(group);
-
-
-        public WatermarkOffsets GetWatermarkOffsets(TopicPartition topicPartition)
-            => consumer.GetWatermarkOffsets(topicPartition);
-
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition, TimeSpan timeout)
-            => consumer.QueryWatermarkOffsets(topicPartition, timeout);
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition)
-            => consumer.QueryWatermarkOffsets(topicPartition);
-
-
-        public Metadata GetMetadata(bool allTopics, TimeSpan timeout)
-            => consumer.GetMetadata(allTopics, timeout);
-
-        public Metadata GetMetadata(bool allTopics)
-            => consumer.GetMetadata(allTopics);
+        public Client ClientInstance
+            => consumer.ClientInstance;
     }
 
     public class Consumer : IDisposable
@@ -493,6 +464,8 @@ namespace Confluent.Kafka
             {
                 throw new KafkaException(pollSetConsumerError, "Failed to redirect the poll queue to consumer_poll queue");
             }
+
+            ClientInstance = new Client(kafkaHandle);
         }
 
         /// <summary>
@@ -817,49 +790,6 @@ namespace Confluent.Kafka
         public long OutQueueLength
             => kafkaHandle.OutQueueLength;
 
-
-        public List<GroupInfo> ListGroups(TimeSpan timeout)
-            => kafkaHandle.ListGroups(timeout.TotalMillisecondsAsInt());
-
-        // TODO: is a version of this with infinite timeout really required? (same question elsewhere)
-        public List<GroupInfo> ListGroups()
-            => kafkaHandle.ListGroups(-1);
-
-
-        public GroupInfo ListGroup(string group, TimeSpan timeout)
-            => kafkaHandle.ListGroup(group, timeout.TotalMillisecondsAsInt());
-
-        public GroupInfo ListGroup(string group)
-            => kafkaHandle.ListGroup(group, -1);
-
-
-        public WatermarkOffsets GetWatermarkOffsets(TopicPartition topicPartition)
-            => kafkaHandle.GetWatermarkOffsets(topicPartition.Topic, topicPartition.Partition);
-
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition, TimeSpan timeout)
-            => kafkaHandle.QueryWatermarkOffsets(topicPartition.Topic, topicPartition.Partition, timeout.TotalMillisecondsAsInt());
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition)
-            => kafkaHandle.QueryWatermarkOffsets(topicPartition.Topic, topicPartition.Partition, -1);
-
-
-        /// <summary>
-        ///
-        /// </summary>
-        /// <param name="allTopics">
-        ///     true - request all topics from cluster
-        ///     false - request only locally known topics (topic_new():ed topics or otherwise locally referenced once, such as consumed topics)
-        /// </param>
-        /// <remarks>
-        ///     TODO: Topic handles are not exposed to users of the library (they are internal to producer).
-        ///           Is it possible to get a topic handle given a topic name?
-        ///           If so, include topic parameter in this method.
-        /// </remaarks>
-        public Metadata GetMetadata(bool allTopics, TimeSpan timeout)
-            => kafkaHandle.GetMetadata(allTopics, null, timeout.TotalMillisecondsAsInt());
-
-        public Metadata GetMetadata(bool allTopics)
-            => kafkaHandle.GetMetadata(allTopics, null, -1);
+        public Client ClientInstance { get; }
     }
 }
