+++ src/Confluent.Kafka/Producer.cs
@@ -280,6 +280,8 @@ namespace Confluent.Kafka
                 callbackCts = new CancellationTokenSource();
                 callbackTask = StartPollTask(callbackCts.Token);
             }
+
+            ClientInstance = new Client(kafkaHandle);
         }
 
         // TODO: think about Poll and flushing.
@@ -404,41 +406,7 @@ namespace Confluent.Kafka
             kafkaHandle.Dispose();
         }
 
-        // TODO: potentially use interface to avoid duplicate method docs.
-
-        public List<GroupInfo> ListGroups(TimeSpan timeout)
-            => kafkaHandle.ListGroups(timeout.TotalMillisecondsAsInt());
-
-        public List<GroupInfo> ListGroups()
-            => kafkaHandle.ListGroups(-1);
-
-
-        public GroupInfo ListGroup(string group, TimeSpan timeout)
-            => kafkaHandle.ListGroup(group, timeout.TotalMillisecondsAsInt());
-
-        public GroupInfo ListGroup(string group)
-            => kafkaHandle.ListGroup(group, -1);
-
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition, TimeSpan timeout)
-            => kafkaHandle.QueryWatermarkOffsets(topicPartition.Topic, topicPartition.Partition, timeout.TotalMillisecondsAsInt());
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition)
-            => kafkaHandle.QueryWatermarkOffsets(topicPartition.Topic, topicPartition.Partition, -1);
-
-
-        public WatermarkOffsets GetWatermarkOffsets(TopicPartition topicPartition)
-            => kafkaHandle.GetWatermarkOffsets(topicPartition.Topic, topicPartition.Partition);
-
-
-        private Metadata GetMetadata(bool allTopics, string topic, int millisecondsTimeout)
-            => kafkaHandle.GetMetadata(allTopics, topic == null ? null : getKafkaTopicHandle(topic), millisecondsTimeout);
-
-        public Metadata GetMetadata(bool allTopics, string topic, TimeSpan timeout)
-            => GetMetadata(allTopics, topic, timeout.TotalMillisecondsAsInt());
-
-        public Metadata GetMetadata(bool allTopics, string topic)
-            => GetMetadata(allTopics, topic, -1);
+        public Client ClientInstance { get; }
     }
 
 
@@ -615,41 +583,7 @@ namespace Confluent.Kafka
         public void Dispose()
             => producer.Dispose();
 
-
-        public List<GroupInfo> ListGroups(TimeSpan timeout)
-            => producer.ListGroups(timeout);
-
-        public List<GroupInfo> ListGroups()
-            => producer.ListGroups();
-
-
-        public GroupInfo ListGroup(string group, TimeSpan timeout)
-            => producer.ListGroup(group, timeout);
-
-        public GroupInfo ListGroup(string group)
-            => producer.ListGroup(group);
-
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition, TimeSpan timeout)
-            => producer.QueryWatermarkOffsets(topicPartition, timeout);
-
-        public WatermarkOffsets QueryWatermarkOffsets(TopicPartition topicPartition)
-            => producer.QueryWatermarkOffsets(topicPartition);
-
-
-        public WatermarkOffsets GetWatermarkOffsets(TopicPartition topicPartition)
-            => producer.GetWatermarkOffsets(topicPartition);
-
-
-        /// <summary>
-        ///     - allTopics=true - request all topics from cluster
-        ///     - allTopics=false, topic=null - request only locally known topics (topic_new():ed topics or otherwise locally referenced once, such as consumed topics)
-        ///     - allTopics=false, topic=valid - request specific topic
-        /// </summary>
-        public Metadata GetMetadata(bool allTopics, string topic, TimeSpan timeout)
-            => producer.GetMetadata(allTopics, topic, timeout);
-
-        public Metadata GetMetadata(bool allTopics, string topic)
-            => producer.GetMetadata(allTopics, topic);
+        public Client ClientInstance
+            => producer.ClientInstance;
     }
 }
