+++ src/Confluent.Kafka/project.json
@@ -14,7 +14,7 @@
     },
 
     "frameworks": {
-        "net451": { },
+/*        "net451": { },*/
         "netstandard1.3": {
             "dependencies": {
                 "System.Console": "4.0.0",
